Fixes into iOS NavigationRenderer to better support animations (#7)
* Add built-in transition handling and update navigation

- Introduced `IsBuiltIn` method in `AnimationHelpers.cs` to check for built-in transition types.
- Updated `BuiltInAnimation` in `AnimationHelpers.ios.cs` to handle animations without adding `currentView` to the `window`.
- Adjusted animation logic in `NavigationTransRenderer.cs` to utilize the new `IsBuiltIn` method.
- Increased transition duration in `NewPage1.xaml.cs` from 500 to 4500 milliseconds.
- Modified tap gesture handler in `NewPage2.xaml.cs` to check for Shell usage before navigating.
- Created new solution file `PJ.NavigationTrans.Pjro.sln` for project structure and configurations.

* Add OnPopViewAsync method to NavigationTransRenderer

Implemented asynchronous view popping with animation in the
NavigationTransRenderer class. Adjusted the insertion index
for currentView in the window to improve view layering
during transitions.

* Refactor navigation animation handling

- Updated `OnPopViewAsync` to return base method result after animation.
- Replaced direct animation logic with `HandlePop()` and `HandlePush()` methods.
- Improved code organization and readability by encapsulating animation logic.

* pop animation is almost back

* Improve view controller popping and remove MemoryTest class

Updated the `NavigationTransRenderer` class to enhance view controller popping behavior by introducing an `ignorePopCall` flag and refining the `PopViewController` method. Simplified the `HandlePop` method by removing extraneous comments and code.

Removed the `MemoryTest` static class and its associated methods, as it was deemed unnecessary for managing weak references to `UIView` objects.

* Swap animation order in HandlePop method

Reversed the order of animation calls in the
`HandlePop` method of the `NavigationTransRenderer` class.
This change updates the visual transition effect when
popping a view in the navigation renderer.

* update yml

* fixed typo

* dump change to trigger CI 🤞🏾

* Add workflow_dispatch and environment variables

This commit adds the `workflow_dispatch` event to both `dotnet-macos.yml` and `dotnet-windows.yml`, enabling manual triggering of workflows. It also defines the environment variables `LIB_CSPROJ` and `SAMPLE_CSPROJ` for project file paths. Additionally, `dotnet-windows.yml` includes a new `NugetPackageVersion` variable to format the NuGet package version with the run number.

* fixed typo